update : ë¡request body 로ê깅 , 커ì텀 익셉션 ì개발

+++ BE/zeepseek/src/main/java/com/zeepseek/backend/domain/property/dto/response/PropertySummaryDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PropertySummaryDto {
-    private Long propertyId;
+    private Integer propertyId;
     private Double latitude;
     private Double longitude;
 }
